backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.0.62
langchain-openai==0.1.25
langchain-core>=0.2.40
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
